backend/requirements.txt
fastapi
agno
openai
pgvector
PyPDF2
psycopg[binary]
sqlalchemy
uvicorn
groq
python-dotenv
duckduckgo-search
playwright
selenium
pytest
pytest-playwright
browserbase
python-docx
python-multipart
